History lookup now checks child if no parent found
Problem expressed in prior commit was found to be an error in mame history file with my test entry - outruna is now outrunra.
Meaning multiline lookups doubtless fine - I was seeing that MameUI is defaulting to searching for parent rom entry in cases where child isn't found.
Which is a good idea...