Multiloader decides unzip policy with symlinks
I decided that moving a zip locally to unzip it is one
use case, another is you want to extract a file as quick as possible.
In that latter case you've probably got the zip local.

I wanted something simple to allow users to specify that if a file is
close, just extract it, but if it is far, copy the zip locally THEN
extract it.

I decided the answer was "if the cd/dvd image is a symlink, it needs
moving locally, if its the file itself, it doesnt.