fixed bug with efind window and large displays
windows users must change dpi for large displays
(andlog out!), this hasn't changed for years, but
fEmuFind.pas' main window was configured with
visual/autosize set to off, so all 3 panes in it
were unusuable.

A to do is to investigate why delphi
complained at first that onClick methods in this .pas
are invalid

I changed the guest OS' dpi to 120 in order to get a native
view of the problem. i left it like that