regreattable exception for nulldc
i noticed some time ago that some non cdi/gdi
images didn't run up if passed not mounted first.
Hence the two loaders. The direct loader has a problem
though in the novel way nulldc does its config and command line
- quite brilliant as the command line directly references config
keys, but not so good when we need to call the name of the rom
as a separate parameter and not as part of another parameter.

This exception needs to, at least, be refactored away nicely