[mame] filepath opts to tab group in mame opts...
due to the visibility and invalidation requirements between the general
mame options and the mame filepath options, it seems much cleaner to
keep them on one form. To make this fit consider that the merge and zip
options for filepaths are both standalone and have cler defaults,
therefore suit the non-default tab.

A main conideration here is: if the Mame emulator changes, we must
invalidate the rompath and the user's rompath -> pathtype selections.
With these displayed on the general mame options form, this becomes
sensible. Downside is that an optional (off by default) series of
options become more prominant