[mame] rompath clearing logic shown workking...
but note a clear is INLINE when form is loaded, this is for
dmeonstration. The salient points are that we set the rompath types to
'' when they aren't set, and we DON'T save the rommpath from the
mame.ini unless you've set a rompath. This is in response to the user
stories written this very day. To get this working we need to
1) stop clearing the settings on each load
2) some way of refreshing the settings on disk that doesn't reset the
   entire world ( again see the user stories - we need to sometimes
   reset even if we haven't pressed 'ok' on the form, but the reset all
   is far too time consuming for this form)