LIB DIR ADDED TO CODEBASE...
previous thinking: since the lib dir is 300 meg of random files,
only commit the lib installs dir - we don't like to commit the
package-managed individual files in our own code repository.
Particularly as some of these files appear to need to be installed
to windows (eg:JVCL).

new thinking: every single person who has interacted with me on the
codebase has benefitted from me zipping up this lib dir and sending it
to them, given the intricacy of installing and linking to the libs
correctly. So include them

The downside here is that we must accept we may need to install over
the top or better still use the lib dir for reference only and
reinstall everything. Also we must not get sloppy for the same reason
and update only live libs without incorporating the changes back to the
sources (particularly john lib, which really should just be included in
the quickplay source for ease of use, given the number of times i've had
to change it)

Ultimately this seems like a necessary, if painful and irksome, step
forward